backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
google-auth
google-auth-oauthlib
google-api-python-client
cryptography
httpx
python-dotenv
alembic
pydantic
